implemented session mngt with filter(s) , created a landing page, implemented delete confirmation on delete action, and proper error ad success handling

+++ build/web/ViewCoursePage.jsp
@@ -77,7 +77,7 @@
 //                                      out.print("<td>"+Courseid+"</td>");
                                       out.print("<td>");
                                       out.print(" <a class='edit' href='CourseUpdate.jsp?id="+ Courseid +"'>Edit</a>");
-                                      out.print("<a class='delete' href='CourseDelete?id="+ Courseid +"'>Delete</a>");
+                                      out.print("<a class='delete' onClick='ConfirmDelete(event)' href='CourseDelete?id="+ Courseid +"'>Delete</a>");
                                       out.print("</td>");
                                       out.print("</tr>");
                                  }
@@ -86,6 +86,27 @@
                                 }
                         
                         %>
+                        
+                         <%
+                    // Check for the presence of the error message
+                    String DeleteMessage = (String) request.getAttribute("DeleteMessage");
+                    String UpdateMessage = (String) request.getAttribute("UpdateMessage");
+                    String SuccessMessage = (String) request.getAttribute("SuccessMessage");
+                    if (DeleteMessage != null) {
+                %>
+                <p style="font-size: 20px; text-align: center; color: white;" id='errMess'><%= DeleteMessage%></p>
+                <%
+                    }else if(SuccessMessage!=null){
+                %>
+                <p style="font-size: 20px; text-align: center; color: white;" id='errMess'><%= SuccessMessage%></p>
+                <%
+                }else if(UpdateMessage!=null){
+                %>
+                <p style="font-size: 20px; text-align: center; color: white;" id='errMess'><%=UpdateMessage %></p>
+                 <%
+                }
+                %>
+                
                                 <!-- Add more rows as needed -->
                                 </tbody>
                             </table>
@@ -97,4 +118,26 @@
         </div>
     </div>
 </body>
+ <script>
+         const Ptag= document.getElementById('errMess');
+            setTimeout(()=>{Ptag.innerHTML="";},10000);
+            console.log(Ptag)
+            
+//                  function ConfirmDelete(event){
+//             if(confirm("Are you sure you want to delete this?")){
+//                 console.log("deleted");
+//             }else{
+//                 alert("Delete Canclled");
+//                 event.preventDefault();
+//             }
+             
+             function ConfirmDelete(event){
+                 if(confirm("Are you sure you want to delete this?")){
+                     console.log("deleted successfully");
+                 }else{
+                     alert("delete action cancelled")
+                     event.preventDefault();
+                 }
+             }
+    </script>
 </html>
