+++ build/web/ViewFacultyPage.jsp
@@ -68,7 +68,7 @@
                                       out.print("<td>"+FacName+"</td>");
                                       out.print("<td>");
                                       out.print(" <a class='edit' href='FacultyUpdate.jsp?id="+ FacID +"'>Edit</a>");
-                                      out.print("<a class='delete' href='FacultyDelete?id="+FacID+"'>Delete</a>");
+                                      out.print("<a class='delete' onClick='ConfirmDelete(event)' href='FacultyDelete?id="+FacID+"'>Delete</a>");
                                       out.print("</td>");
                                       out.print("</tr>");
                                  }
@@ -76,7 +76,27 @@
                                 System.out.println(e);
                             }
 
-                        %>                
+                        %>  
+                         <%
+                    // Check for the presence of the error message
+                    String DeleteMessage = (String) request.getAttribute("DeleteMessage");
+                    String UpdateMessage = (String) request.getAttribute("UpdateMessage");
+                    String SuccessMessage = (String) request.getAttribute("SuccessMessage");
+                    if (DeleteMessage != null) {
+                %>
+                <p style="font-size: 20px; text-align: center; color: white;" id='errMess'><%= DeleteMessage%></p>
+                <%
+                    }else if(SuccessMessage!=null){
+                %>
+                <p style="font-size: 20px; text-align: center; color: white;" id='errMess'><%= SuccessMessage%></p>
+                <%
+                }else if(UpdateMessage!=null){
+                %>
+                <p style="font-size: 20px; text-align: center; color: white;" id='errMess'><%=UpdateMessage %></p>
+                 <%
+                }
+                %>
+                
                                 <!-- Add more rows as needed -->
                                 </tbody>
                             </table>
@@ -88,4 +108,19 @@
         </div>
     </div>
 </body>
+ <script>
+         const Ptag= document.getElementById('errMess');
+            setTimeout(()=>{Ptag.innerHTML="";},10000);
+            console.log(Ptag)
+            
+            function ConfirmDelete(event){
+                if(confirm("are you sure u want to delete this?")){
+                    console.log("deleted");
+                }
+                else{
+                    alert("Delete action cancelled");
+                    event.preventDefault();
+                }
+            }
+    </script>
 </html>
