+++ web/ViewPayment.jsp
@@ -84,7 +84,7 @@
                                       out.print("<td>"+StudID+"</td>");
                                       out.print("<td>");
                                       out.print(" <a class='edit' href='PaymentUpdate.jsp?id="+ StudID +"'>Edit</a>");
-                                      out.print("<a class='delete' href='PaymentDelete?id="+StudID+"'>Delete</a>");
+                                      out.print("<a class='delete' onClick='ConfirmDelete(event)' href='PaymentDelete?id="+StudID+"'>Delete</a>");
                                       out.print("</td>");
                                       out.print("</tr>");
                                  }
@@ -92,7 +92,28 @@
                                 System.out.println(e);
                             }
 
-                        %>                
+                        %>   
+                        
+                         <%
+                    // Check for the presence of the error message
+                    String DeleteMessage = (String) request.getAttribute("DeleteMessage");
+                    String UpdateMessage = (String) request.getAttribute("UpdateMessage");
+                    String SuccessMessage = (String) request.getAttribute("SuccessMessage");
+                    if (DeleteMessage != null) {
+                %>
+                <p style="font-size: 20px; text-align: center; color: white;" id='errMess'><%= DeleteMessage%></p>
+                <%
+                    }else if(SuccessMessage!=null){
+                %>
+                <p style="font-size: 20px; text-align: center; color: white;" id='errMess'><%= SuccessMessage%></p>
+                <%
+                }else if(UpdateMessage!=null){
+                %>
+                <p style="font-size: 20px; text-align: center; color: white;" id='errMess'><%=UpdateMessage %></p>
+                 <%
+                }
+                %>
+                
                                 <!-- Add more rows as needed -->
                                 </tbody>
                             </table>
@@ -104,5 +125,19 @@
         </div>
     </div>
     </body>
+     <script>
+         const Ptag= document.getElementById('errMess');
+            setTimeout(()=>{Ptag.innerHTML="";},10000);
+            console.log(Ptag);
+            
+            function ConfirmDelete(event){
+                if(confirm("are you sure you want to delete this?")){
+                    console.log("deleted");
+                }else{
+                    alert("delete action cancelled");
+                    event.preventDefault();
+                }
+            }
+    </script>
 </html>
