+++ web/ViewStudents.jsp
@@ -80,7 +80,7 @@
                                       out.print("<td>"+Reg+"</td>");
                                       out.print("<td>");
                                       out.print(" <a class='edit' href='StudentUpdate.jsp?id="+ StudentID +"'>Edit</a>");
-                                      out.print("<a class='delete' href='StudentsDelete?id="+StudentID+"'>Delete</a>");
+                                      out.print("<a class='delete' onClick='ConfirmDelete(event)' href='StudentsDelete?id="+StudentID+"'>Delete</a>");
                                       out.print("</td>");
                                       out.print("</tr>");
                                  }
@@ -89,20 +89,26 @@
                             }
 
                         %>                
-                <%
+             <%
                     // Check for the presence of the error message
-                    String errorMessage = (String) request.getAttribute("ErrorMessage");
-                    String Message = (String) request.getAttribute("SuccessMessage");
-                    if (errorMessage != null) {
+                    String DeleteMessage = (String) request.getAttribute("DeleteMessage");
+                    String UpdateMessage = (String) request.getAttribute("UpdateMessage");
+                    String SuccessMessage = (String) request.getAttribute("SuccessMessage");
+                    if (DeleteMessage != null) {
                 %>
-                <p style="font-size: 20px; text-align: center; color: white;" id='errMess'><%= errorMessage%></p>
+                <p style="font-size: 20px; text-align: center; color: white;" id='errMess'><%= DeleteMessage%></p>
                 <%
-                    }else if(Message!=null){
+                    }else if(SuccessMessage!=null){
                 %>
-                <p style="font-size: 20px; text-align: center; color: white;" id='errMess'><%= Message%></p>
+                <p style="font-size: 20px; text-align: center; color: white;" id='errMess'><%= SuccessMessage%></p>
                 <%
+                }else if(UpdateMessage!=null){
+                %>
+                <p style="font-size: 20px; text-align: center; color: white;" id='errMess'><%=UpdateMessage %></p>
+                 <%
                 }
                 %>
+                
                                 <!-- Add more rows as needed -->
                                 </tbody>
                             </table>
@@ -115,8 +121,17 @@
     </div>
     <script>
          const Ptag= document.getElementById('errMess');
-            setTimeout(()=>{Ptag.innerHTML="";},4000);
+            setTimeout(()=>{Ptag.innerHTML="";},10000);
             console.log(Ptag)
+
+            function ConfirmDelete(event){
+                if(confirm("Are you sure you want to delete this?")){
+                    console.log("deletd");
+                }else{
+                    alert("Delete action cancelled");
+                    event.preventDefault();
+                }
+            }
     </script>
 </body>
 </html>
